Description: modify the code problems IssueNo: https://gitee.com/openharmony/graphic_ui/issues/I5001O Feature or Bugfix: Bugfix Binary Source:No

+++ frameworks/src/core/base/dfx_assist.cpp
@@ -97,7 +97,7 @@ void DfxAssist::DumpErrorMessage(const jerry_value_t errorValue)
         return;
     }
     HILOG_ERROR(HILOG_MODULE_ACE, "[Exception backtrace]:");
-    for (uint8_t i = 0; i < length; i++) {
+    for (uint32_t i = 0; i < length; i++) {
         jerry_value_t itemVal = jerry_get_property_by_index(backtraceVal, i);
         jerry_size_t strSize = 0;
         if (!jerry_value_is_error(itemVal) && jerry_value_is_string(itemVal)) {
